Split IRemoteGitRepo from IGitRepo
